Fix compilation on Linux
